+++ local/sharim.py
@@ -6,7 +6,7 @@ from requests.exceptions import Timeout
 import json
 
 def sendh(data,url):
-    
+    pass
 
 global round_number
